Add quick dependency fix for djblets to setup.sh.
'pip install' doesn't work for djblets package, so it cannot be given
in pip-requires. Also, i've made some changes to djblets templates, so
it is obligatory to use forked repository in easy_install dependencies.
As an alternative, forked djblets sources can be downloaded directly
from git repository, unpacked and installed via setup.py install - all this
done inside setup.sh script. It is surely not the most proper way, but
is the fastest one.

Change-Id: Ibfbb2913ce4a31911ef5415deea99e6f99926c2e