Remove djblets package from murano-dashboard.
Use horizon.tables package instead for drawing the tables (old term is
datagrid).  Now columns attribute is configurable in Service
YAML-markup (it should be consistent with initial data for
table). Last thing to be done is to somehow employ Actions from
horizon.tables (if this can be done at all) - this eliminates need in
custom js-code which adds and deletes table rows on client-side.

Change-Id: I6e6cd41b717d48eafd7e4a5e1fc907d2f1ebf12d
Closes-bug: #1234194