Remove service selection step from Create Service (App) Wizard.
This simplification is aimed to:
 * integrate Create App Wizard with the new App Catalog where the App
   to be created is selected before the Wizard is instantiated
 * allow 'Add App' button logic for composite fields (see mentioned
   blueprint) during creation of other Application

As part of this simplification metadata cache is temporarily dropped
at all - so the first ui definition received from the metadata service
will ever be used until web-server restart. Actually instead of real
metadata service apiary mockup is used for now.

Also, as part of this change the 'Create Service' button was removed
from the Environment->Services table in favor of 'Add Application'
button (which redirects user to the App Catalog page from which he
actually adds new Apps to that Environment).

Partially implements: blueprint new-web-ui-prototype
Change-Id: I85d4cf4a150cce781abf6473e493533434f76ddd