Fill 'Latest Apps' section with recently deployed/uploaded Apps
Each App that was recently added to an Environment using Deploy/Quick
deploy button (actually, just pressing that button is enough) or
uploaded to catalog at Packages panel is added to Latest Apps list on
top of AppCatalog. To refresh that list with newly deployed App, on
final step Wizard is redirected to AppCatalog panel instead of simply
closing modal form. This fixes bug #1306579 as well.

That commit is dependent on https://review.openstack.org/#/c/90315/
and should be merged once that change gets into
python-muranoclient>=0.5.1.

Change-Id: Iefe2abb9efb5ae8cbe2d26809eb876eb0ecd5906
Closes-Bug: #1312213
Closes-Bug: #1306579