Enable passing several FQNs as one field's type
These types should be specified as string (no YAML sequences!), string
entities are separated with ', '.  It is a kind of workaround until
descendants of some abstract MuranoPL class given by its FQN cannot be
calculated by murano-api - so we have to provide them explicitly. But
even after that issue is fixed, the client-side code can be reused.

Change-Id: I0cbf26b8fd6834378943a600abf98372d45f1788