Fix extra column in services table during deploy
After clicking deploy environment button, the services table will
randomly show a checkbox while it updates the row data via ajax, even
though the header column for the multiselect checkbox is hidden. It
seems to be a bug in horizon.tables.js that is preventing the 'hidden'
attribute of the multi_select_column data cell to be honored. The
workaround will be to remove the checkboxes and to remove the
'Delete Services' button at the top of the table. Services can still
be deleted on a row-by-row basis in the interim.

Change-Id: I5451e9bcfb8dd9c9650cea7a10210891f2200e57
Closes-Bug: 1335339