Add color distinction for deployed and non-deployed VMs
This patch adds a color distinction between already deployed instances
and on deployed. It is good option for autoscaling App representation
as topology page shows maximum of instances in the cluster but color
allows to understand what is actually deployed.
Deployed Vms are green and non deployed are gray.
Also resize topology page for a full view.

Change-Id: Ie8024b86b9e9a9a88dda77075e3fc8bc81e1a06f
Partly-Implements: bp/topology-page