Make category selection a separate step
After package file is upload, user should point to witch category this package belongs.
Now, it's not obvious that this parameter is requred and user don't undestand
what he needs to do. Making a separate step for category selection will make that clear.
Also, templates for that forms were updated.

Closes-Bug: #1394572
Change-Id: Iec964a5b09602b465e76a537344d3a900478cc96