Fix horizon dependency for stable/juno
Dashboard is built on top of Horizon, but does not state Horizon
as a requirement in requirements.txt. Instead it is done right
in the tox.ini file. In most of the production cases when Dashboard
is installed as plugin for the OpenStack Dashboard Horizon is already
installed therefore we need this requirement only for development
purposes.

Horizon is referenced as a link to tarball in tox.ini, but in stable/juno
branch wrong link was referenced, horizon-master.tar.gz instead of
horizon-2014.2.tar.gz.

Correct version for Horizon in the next release of Dashboard from stable/juno
is horizon-2014.2.1.tar.gz

Change-Id: Ic9179fe880a42e18f1c9e3321385937c6dfbc3ac
Closes-bug: #1412164