Tri-state checkbox widget
Tri-state checkbox is the kind of checkbox which has three states
instead of two, as the regular checkbox has.

Third state is "indeterminate", which means, that the element is in
undefined state. Common example of such situation is some parent
element, which has some of child elements selected and some not - and
so it means that parent element is not selected nor unselect - which
in terminology of this changeset called "indeterminate" state.

Partially implements bp assign-category-button
Depends-On: Ie65a77eab29aa191420fbd630b1d242b59c9ff9f
Change-Id: I98116fce362af5eb4c64941f4c442dd81120e762