Environments: reload page after all the environments have been updated
Before murano had an overriden datatable script in
horizon.tables+reload.js. This script was copy-pasted for a single
single line edit. This is actually not required, because horion
datatables fire a "update" trigger after row updates,
which is enough for the purpose of updating the page.

Before any row in the ready/deploy_failure would trigger a reload.
With this patch only an update trigger is set for the services table
instead of overriding whole datatable js routine. The page is
reloaded only after there are no rows with 'status_unknown', i.e.
when every app either deployed or failed to deploy.

Change-Id: I254538cb776d2dee422be0363a736a5a9c1c9e2b
Closes-Bug: #1454221