Mark BooleanFields as non-required in dynamic_ui
Due to the way django validates fields having a required=True fieild
almoust never makes sense. This would mean that user would not be able
to choose false for the boolean value, that the Field represents,
because django would raise an error, stating, that the field is
required.
This is especially true for fields, generated for HOT template-packages.
This patch sets required to False for dynamic_ui.

Change-Id: Ibc36017e37bb7df9f7bde2c53265237b21276fd6
Related-Bug: #1450536