Remove unnecessary string conversions for status reports
Status.text is already a string or a unicode object. That is because
text is of type sa.Text() in database, and api serializes it as such.

Unnecessary conversion to str prevented unicode statuses from displaying
correctly in dashboard (although deployment continued normally).

This change allows murano dashboard to display unicode statuses. We ♥
unicode.

Change-Id: Ic02d2d55c13a555fd878f50b83f32a4f36a6904f
Related-Bug: #1361169