Escape app names, returned to horizon modal forms
When adding an app to the environment via plus button on the form of
another app an ajax request is made by horizon js code, to get the name
of the app, that was just added to env. Name of the app can contain
malicious code (alternatively the code can be present in the initial
value of the name).


This patch escapes name, returned during that request.

Change-Id: I31f63ed1aa0d46496978110f45c7b43a04519970
Closes-Bug: #1460493