Remove attempts to close success message in tests
This commit removes attempts to close a success message, because horizon code
may hide and remove the message before tests are able to close it. This
could trigger false negatives from time to time.
It's harmless to ignore success the message as long as it is present on the page.

Change-Id: I660d174493cade75fcab6a8f5b7f6214dd188589
Related-Bug: #1466332