Remove unused and unsupported table field
This commit removes any mention of table field from dynamic_ui and
removes floppyforms from requirements.

'table' field was introduced for MsSqlClusterServer service, that
changed significantly since then and does not use table field anymore.
See
https://github.com/murano-project/murano-app-incubator/blob/master/io.murano.windows.SqlServerCluster/UI/ui.yaml
for more details (even though the repository is deprecated).

Since no app available today uses this type of field, and no tests exist
for this type of field. There is no way to test, that it works correctly,
therefore it is very hard to transfer critical parts from floppyforms to
murano-dashboard.

'table' field depends on django-floppyforms, which is not
present in global-requirements, which is problematic and prevents us
from enabling proposal bot in murano-dashboard.

'table' field was never documented, so no documentation change is
required.

Change-Id: I638c722c2bdde411aaefa02f3d4027119de0e0bd
Closes-Bug: #1468342