Compress murano-dashboard related js in a single file
this commit introduces 'murano_base.html' as the base template for all
murano-dashboard templates. This commit contains js block, that contains
and compresses all of the murano-dashboard js code. All js are moved
from individual pages to murano_base.html. This means, that
murano-dashboard js code has to behave nice, when put on a page, not
designed for it and should not assume, that certain elements are present
in DOM tree.
To achieve this the commit does the following:
* makes selectors more specific for 'environments-in-place.js'
* makes topology code only launch on tab init
* makes import form code only launch on modal init

Also moves modal init code to ADD_JS_FILES section, to allow calling
murano modals from external dashboards (i.e. app-catalog-ui)

Change-Id: I342702b6b236b3323f57a4a26fdf4120505b9800
Partial-Bug: #1358233