Add eslint javascript checker
This commit adds a eslint tox environment, that installs eslint and
checks muranodashboard js code against it.

Eslint is the current linter of choice for horizon, because it is
capable of detecting both style and syntax errors (and does not have
license limitations of jshint).

To run: tox -e eslint
Partially implements: blueprint add-js-lint-jobs

Change-Id: I645f279ab6ac45b7d1d872ded803c126dd0a10e7