Catalog: show package_count for categories.
Filtering by a category, that does not contain any packages would result
in an empty set of packages. We can prevent user from doing so, by
disaplying number of packages in a category.
This commit changes packages.categories to categories.list, that
includes package_count information and outputs the information in the
dropdown menu.
This commit also sorts categories so that categories with packages
are listed before categories without packages
(both groups alphabetically ordered) for better ux.

Change-Id: Ieaa0d64cd4bc703f4918fa41cb280288f76c2aa2
Related-Bug: #1454865