Add NetworkChoiceField to Dynamic UI
Sometimes a VM should be placed to an existing network rather then to a
new network created during deployment. While our workflows support this,
there is no way for the end user to select such a network in the UI.

This patch adds a new type of the field for Murano Dynamic UI:
NetworkChoiseField. This field provides a selection of networks and
their subnetworks as a dropdown pupulated with those which are available
to the current tenant.

When the user selects an item in this list the IDs of selected network
and subnet are returned by the control and thus may passed to the
engine along with the rest of the ObjectModel, so the newly created
instances join the specified net and subnet.

The control has a default option ("Auto"): if it is selected the
returned tuple contains Null, so the ObjectModel may be constructed in a
default way. This option may be disabled by the field parameter, so
package developers may use this field to oblige the user to select some
network.

Field has some other parameters, which allow to filter the available
networks: pick only the networks matching the specific regex, exclude
murano-generated networks or replace murano-generated names with
human-readable ones.

As this new field adds new functionality to the Dynamic UI platform,
it also bumps up a minor version of the format version to 2.1.

Implements Blueprint ui-network-selection
DocImpact

Change-Id: Ie256c36523c67f27e4cb0b7c9aeed76b2abd28af