Improve the checks of Dynamic UI format versions
Current implementation of DynamicUI checks the "Version" field of UI
definition files to exactly match a single constant value (currently 2).
This contradicts to our desired versioning strategy.

Instead, it should check that the major component of the Version field
in the UI definition being processed  matches tha major component of the
current latest version of the format (i.e. we assume that the versions
are compatible if their major version are the same). At the same time
the version of the definition being processed is not greater than the
latest supported version.

The implementation is done using the semantic_version library which has
all the needed functionality, including the coercing method (parsing of
string-based content into the proper version) and partial versions.

Targets Blueprint murano-versioning

Change-Id: I96cbbddc4389804fafe9a1d6f48f63ac4ee5620a