Allow to select default Environment Network
Provides users with an ability to select an existing Neutron network to
be the default default network of the given environment. Thus all the
VM instances running applications of this environment will join that
network unless individually configured to join some other net.
This allows to override the default behavior, i.e. the creation of
dedicated network+subnet combination for each environment.

This is imlemented by modifying the CreateEvironmentForm dialog: it's
got a new dropdown field which allows to pick the network to join. This
dropdown shares the same logic with NetworkChoiceField from Dynamic UI.
By default, this dropdown has a "Create New" value selected, thus
falling back to old behavior.

Implements blueprint: ui-network-selection
DocImpact

Change-Id: I20e5d0ee498a1d67dd836fb27d6b1200cd0dd11f