Stringify LATEST_FORMAT_VERSION in dynamic_ui
The 'latest version' was initially specified as integer and used direct
comparison. The introduction of semver notation in I96cbb added second
segment to the version, i.e. it became X.Y. As no quotes were added this
became a floting point number. Semver function was stringifying it
before corcing, so it worked fine, however defining a version as a
floating point is definetely wrong.

This patch stringified the vesion where it is defined and adds a
function which returns a semantic_version.Version object based on it.

Change-Id: Iaed1ecdb25c49af8cf1fca2a555e6d534a380d5a