Choose an approppriate image upon stack failure
The get_app_image function was always choosing
'dashboard/img/stack-green.svg', regardless of the actual status of the
stack. Adding some logic to choose green, red or gray depending on the
status. Also, set as default the 'status' argument to None, since this
function is used on another completely different context

Change-Id: Ifab532ee4706038c85f76061eca85238d0d19be1
Closes-Bug: #1467246