Added the support of Glance Artifact Repository
Murano's application packages can now be stored either in database
(using murano-api) or in Glance Artifact repository (so-called Glance
V3 api). This patch adds an experimental support of the latter
approach to murano dashboard.

As all the difference between these two storages is incapsulated in
the murano client, dashboard just needs to configure it properly by
passing an instance of glance v3 client on creation. This is
controlled by the presence of MURANO_USE_GLARE setting in
local_settings.py. If it exists and is True, muranoclient will store
the packages in Glance. The old mode will be used otherwise.

This patch also adds a "Version" column to the "Package Definitions"
view to display the version of each package. The values are set only
if Artifacts-based storage s in use; the column is populated by dashes
otherwise.

Implements-blueprint: artifact-repository-support
Depends-on: I35d71bb7c03b23b01722e060c3ef770a24168ada
Change-Id: I7f77650436ce4c4845ccbbbabf58872a41d3c14c