Fix initial value for the 'name' field
'name' field is inserted automatically if ui definition has 2.2
version of higher.
If ui definitions were nested, name for parent app was used.
This happend since list of filed were not cleaned up from the previous time.

Also, some improvement was made: initial value is not get from the application section,
but is passed from the fqn, which is get from api.

Change-Id: I01c51d48185400f34a7dbd15fca349ad7ddbef05
Closes-Bug: #1493075