Provide tenant name information
Currently if there are several packages with same name
from different tenant, it's impossible to recognize whitch package
belongs to which tenant.
This commit adds new field to Package Definitions table, that
provides this information.

Non-admin user has no rights to get this information. so this field is
hidden. But it's not a problem, since regular user
is not allowed to modify foriegn packages.
Corresponding tests were updated and refactored a little bit
to keep column order in one place.

Change-Id: Iec796c7eff3d53a1dd34e134aca7ceb3c036c037
Closes-Bug: #1488448