Import images publicly when package is public
When importing a package using package import wizard on the dashboard,
required glance images are imported before having chosen the
visibility of the package (is-public). With this commit, when the
package is set "public" in the final step of the wizard, the glance
images are also set "public".

Closes-Bug: #1507139
Depends−On: I9460360e84cc40d7414aa5cd9f7367f5a80b8466

Change-Id: I64e0572f4986919593b4cf40dac178fcbee9bf6f