Ensure closing of firefox window if setUp fails
In the functional integration test, test class opened a firefox
window as part of setUp() but in case of failure of setUp(), the
window remained open, as the tearDown() method is not called on
failure of setUp(). This problem can be solved by adding the
close command for firefox through addCleanup method which is
excecuted irrespective of success of the setUp method.

Change-Id: Ic576fa05b62afa08356f11825cbf442baaf9a4cb
Closes-bug: 1487446