Add 'password' to names of password fields in dynamic forms
Password-checking javascript selects password fields relying on a
fact that their ids should end with 'password' and 'password-
clone'. But it's not actually true when in application's 'ui.yaml'
file another name for password field is given.
This change adds 'password' to every password field name and id,
and also 'custom_valid' to fields with custom validators to avoid
showing standard warning messages for them.

Change-Id: I4974c221f449b70a46f30e78bfd3cd1b72eeb919
Closes-bug: #1534642