Added a warning message if environment name is empty
Adding required=False to the environment name form allows it to call
update_cell when it has been left empty. Added logic to update_cell
to check if the form is empty or contains only whitespace and issue a
warning if true and only if the form in question is called 'name'.

Change-Id: I60eb475574be7713807fd0f5548a64b1f8c890c7
Closes-Bug: #1571663