Fetch class descendants in the dynamic_ui from glare
Before this change whenever a murano-class dependency was specified in
the ui definition of a package, only packages that contained those
specific fqns were fetched from API, even though the information about
inheritance is available in the glare back-end.

With this change dashboard also fetches all the packages, that have
classes, that inherit from the specified package. This change also
alters the way absence of a package is handled. It's now specified as an
empty field with explanation, rather then a django message. This allows
having only partial dependencies and skipping optional dependencies
altogether

Change-Id: I6632f72ca260c26a116b3aa4608d9592d0f1cbce
Closes-Bug: #1585812