Correct password-checking js in dynamic UI forms
Password-checking javascript selects password fields relying on a
fact that their ids should end with 'password' and 'password-
clone'. But it's not actually true when in application's 'ui.yaml'
file another name for password field is given.
This patch changes js in a way that it doesn't check presence of
'password' in the field id and uses just it's type. Also it
distinguish original password field and confirmation field by
'-clone' ending in id.

Change-Id: I9b9d7cdea8f5c5249cd7595b104424cf7a4055d4
Closes-bug: #1534642