Correct domain parameter value in tests
Previously in tests keystone client method for user creation was called
with domain equal to 'Default', as far as it was supposed to refer to
domain name. However, it really refers to domain id, so value should be
replaced with 'default'.

Closes-Bug:#1603431

Change-Id: Ic7693d74a0ea1a9cdd789d5adc61b07516ac5918
(cherry picked from commit 36355188790212c2ca5365911a0865e93930ccea)