Correct env status in the new session
In the dashboard it is possible to make some changes in the current
session, then close this session and open another one. The result
of this was that changes were not saved but env's status was
changed.
With this patch status of environment is 'Ready to deploy' only
in case when current services list is not equal to the list of
last deployed services.
UI tests to check this behaviour are added.

Change-Id: I925e1f1caf0795960b56cc88b887430dc4871742
Closes-bug: #1606891