Add ability to delete arbitrary component in functional tests
Patch adds no-arguments mode to delete_component() method. It means
that if no component_name is provided as an argument, an arbitrary
component of the env will be deleted. For most test cases the exact
component to delete doesn't matter. It allows to get rid of finding
component's id by its name and thus simplify the action a bit.
Hopefully it will make tests that use this method more reliable.

Change-Id: If32a8691701349a173522fed22e88f40e21784cd
Related-bug: #1618118