Ability to reference single object several times
This commit introduces ref(templateName [, parameterName] [, idOnly])
YAQL function was added to UI definition DSL. This function evaluates
template templateName and fixes the result in parameters under
parameterName key (or templateName if the second parameter was omitted).
Then it generates object ID and places it into ?/id field. On the first
use of parameterName or if idOnly is false the function will return
the whole object structure. On subsequent calls or if idOnly is true it
will return the ID that was generated upon the first call.
Thus the function brings ability to reference single object several
times.

ID generation is done in a safe way with a special temporary type that
wraps string ID and replaced with string representation upon
serialization. This way it is still impossible for a developer to
explicitly provide ID value in the template (it is still going to be
overwritten because it is not of an ObjectID type and it is impossible
to create instance of ObjectID from yaql expression)

Change-Id: Idf114113a73d00ab2132bd085157ca01e94dae3e