Fixes TypeError/other bugs in Environment OverviewTab find_stack.
Recursive calls to find_stack throw TypError because the
parameter `name` is not passed.

In addition, even if `name` is passed, subsequent calls to
find_stack might result in recursion errors, since
subsequent recursive calls to heat_api.stacks_list do not
change any of the arguments passed to that function.
For recursive calls, `marker` should be passed to
heat_api.stacks_list and it should point to the last stack
id returned, so that future calls exclude already-seen
stack results.

From the heatclient.v1.stacks documentation:
:param marker: begin returning stacks that appear later in the stack
               list than that represented by this stack id

Lastly, this bug fix addresses a dictionary key bug with
get_instance_and_stack in the same module. The key
'instances' is mis-typed as 'instance' which is a bug.

Change-Id: Ifd5d4912eb93c7375ae65f1c1bd12eee78959a67
Closes-Bug: #1641422