Topology viewer now properly parses composite type names
In one of previous releases 'type' fields of object model were
modified to hold composite typenames of objects in form of
'fully.qualified.type.name/version@fully.qualified.package.name'.

The topology viewer of Murano Dashboard was not aware of this change
and used the whole value of the type field as a package fqn to
retrieve the application icon from package catalog. This was resulting
in inability find the appropriate package, thus the topology viewer
was always displaying a default icon ("stack") instead of application
icon.

This has been addressed.

Change-Id: I7903cf0b6e554eeea8ce8246811b480fb1881bd6
Closes-bug: #1645367