Fixes unit tests throwing errors, related to Horizon function.
Currently, 11 unit tests are failing because Horizon
changed some of its behavior related to how pagesizes are
calculated from request objects.

This is the commit in Horizon that appears to have caused things
to break in these unit tests:
https://review.openstack.org/#/c/155939/13/horizon/utils/functions.py

This fix mocks out the function in Horizon that does the pagesize
calculation.

Change-Id: I983e83033b8eab25f79f505d2e57ef1e6b86873b
Closes-Bug: #1650029