Fixes murano-dashboard-coverage failing unit test.
environments.test_tables.TestServicesTable.test_get_row_actions
throws an error when run in murano-dashboard-coverage-ubuntu-xenial.

The error happens because get_row_actions in the function under
testing contains an inner class that calls a Horizon function
that performs an API call, throwing the error:
'The default panel "environments" is not registered'.

This mocks out the Horizon function altogther, getting rid of the
error.

Change-Id: Id76b86f97024edb15fd9f093899649bd92fdd680