Add UI Test for Correct # Apps by Category in Catalog View.
Added UI test for the scenario:
"Show correct number of apps in a category in catalog view".

Also made the following 2 tests atomic:
  1) test_add_pkg_to_category_non_admin
  2) test_add_pkg_to_category_without_count
They were not cleaning up after themselves; they were changing
class-level package categories, which was causing the new
test to fail, when run in conjunction with the 2 tests above.

Change-Id: Iac9ba8ed168f4a8cf9144aca285159f66ccea165
Partial-Bug: #1596196