Changes image creation/deletion in sanity_check to be atomic..
Right now, image creation/deletion is done at the class level in
ImageTestCase. Also, in TestSuiteImage relies on self.repair_image
to clean up the class-level image through the UI. This has
several potential points of failure:

 - self.repair_image is not called via self.addCleanup:
   if the test fails before repair_image is called, the image
   is not restored.
 - self.repair_image is done via UI: this can obviously fail,
   because Selenium should never be used to reliably clean up
   a resource.
 - self.image.id is used in many different classes that inherit
   from ImageTestCase: this can result in failures in other
   test suites.

This patch changes self.image to be test-level, created in
setUp instead of setUpClass and deleted in tearDown instead
of tearDownClass.

This is an example of a log that shows multiple tests failing
across different test suites because the class-level image
(probably) was already deleted and thus not found:

https://murano-ci.mirantis.com/logs/25/424525/1/check/gate-murano-dashboard-ubuntu/ca45725/

Change-Id: I0c8628c7fdf182cc512c16faa30f1dd763b8f08e