Fix click issue with test_env_status_new_session_remove_from_two.
The test in the title is failing in the log [1] because, after
deploying the environment with 2 apps, Selenium seems to struggle
interacting with 'Delete Component' button.

This patch simply refreshes the page before clicking on
'Delete Component'.

Also added various explicit waits (which does not reduce
performance much at all) but increases compatibility with slower
machines.

[1] https://murano-ci.mirantis.com/logs/41/425341/1/check/gate-murano-client-ubuntu/34b0037/

Change-Id: Ia6590bcd7eb6e576ee0e82e0771c746ae301dd02