Makes test_delete_multiple_images ui test more stable.
Currently, test_delete_multiple_images in TestSuiteImages fails
irregularly but somewhat frequently. The reason is that, after
calling go_to_submenu to refresh the page with new images, the
new images are not immediately loaded in the table, so clicking
the topmost checkbox does nothing.

This patch splits delete_multiple_images into 2 tests,
one for deleting all the images one by one and one for
deleting all the images at the same time. In the latter case,
each table row is iterated over and its checbox checked, in
order to wait until the row has loaded completely.

It also renames the self.image_title in the base image class to
use 'default' as a prefix, to make debugging failed logs easier.

Change-Id: I523b690144678a3c07afe00786602b24405ffbd0