Fix missing import package button in murano dashboard
After https://review.openstack.org/#/c/476273/ was merged, murano
dashboard is throwing errors related to listing categories when
trying to import a package. Currently, python-muranoclient still
contains a function [0] for listing categories using the deprecated
API v1/packages/categories. The problem is that murano dashboard
calls this function [1][2][3] in python-muranoclient which throws a 404,
as the deprecated API was removed from murano.

This commit replaces instances of client.packages.categories() with
client.categories.list(). However, while the former returns a list
of category names, the latter returns of a list of category objects
-- so they then need to be iterated over to retrieve the list
of category names.

[0] https://github.com/openstack/python-muranoclient/blob/08411aa8d20993ac7c4a52b2aa0e73fb6fea4d40/muranoclient/v1/packages.py#L52
[1] https://github.com/openstack/murano-dashboard/blob/f98e08d827a006db5fd054ac4fb6abba786bb414/muranodashboard/packages/forms.py#L216
[2] https://github.com/openstack/murano-dashboard/blob/f98e08d827a006db5fd054ac4fb6abba786bb414/muranodashboard/packages/forms.py#L288
[3] https://github.com/openstack/murano-dashboard/blob/bc8e4b97d083bfccf916e5538fcdb6ee1cbc4405/muranodashboard/packages/tables.py#L54

Change-Id: I715b1c9ccfdd044980596cc6d966062e0386884c
Partial-Bug: #1701067