Fix adding dependency package to environment with debreach
When using the debreach random comment middleware with Django,
the API response when adding a dependency package to an
environment would end up with a html-style comment at the end of
the response, which would then fail JSON parsing by the dashboard.

To prevent debreach from adding the random data, the content type
needs to be something other than text/html. The response output
itself isn't really JSON so switching to JsonResponse doesn't
work becuase the content isn't a dict, so just use text/plain
instead.

Change-Id: Ie710487d865344c9a5cd71d4dec3bf927cfabf40