Remove 'add' button for keypairs on launch
The add keypair button referenced a view that it no longer
available when using the new angular keypairs panel, otherwise
a 'Danger' error is thrown then you try and click it.

This commit simply removes the add button as there is no way to
get the functionality back when using the angular panel.

Change-Id: Iaf7aa5f0805ba91d2f7f9115cc3cf19cf344dd4b