Drop django-babel from requirement.txt
We have switched the babel extractor for Django from django-babel
to enmerkar recently. murano-dashboard requirements.txt needs to
sync the change. enmerkar, a successor of django-babel will be
installed via horizon, so we don't need to have it in requirements.txt
explicitly.

lower-constrants.txt is updated to use enmerkar accordingly.

Change-Id: Iab0a50a6f6546d3d67d0e9b00f9bbd04ee2f6c7c