Cleanup for lower-constraints job
The patch 7f29364734 removes the lower-constraints job from
the gate but related file like lower-constarints.txt are not
removed. This patch removes the lower-constraints.txt file and
its reference from tox.ini.

Also TC passed a resolution about dropping the lower-constraints.txt
file and its testing. We will keep lower bounds in the requirements.txt
file but with a note that these are not tested lower bounds and we
try our best to keep them updated. For more info please refer [1], [2].

[1] https://etherpad.opendev.org/p/tc-zed-ptg#L326
[2] https://governance.openstack.org/tc/resolutions/20220414-drop-lower-constraints.html#proposal

Change-Id: Ica7454c88d17874ed29c5d29d399ba5684f7f9d1